+++ applications/example/config/files.js
@@ -1,6 +1,6 @@
-// Static files configuration
+{
+  // Static files configuration
 
-module.exports = {
   index: true, // displaye HTTP directory index for /static
   cacheSize: '50mb', // memory cache size
   cacheMaxFileSize: '10mb', // max file size to cache
@@ -9,4 +9,4 @@ module.exports = {
     // 'js', // minify js
     // 'sass' // compile sass to css
   ]
-};
+}