+++ config/sandbox/sandbox.test.js
@@ -1,12 +1,13 @@
 module.exports = {
   api: [
     // Node internal modules
-    'console', 'os', 'fs', 'tls','net', 'dns', 'url', 'util', 'path', 'zlib', 'http', 'https', 'dgram',
-    'stream', 'buffer', 'domain', 'crypto', 'events', 'punycode', 'readline', 'querystring', 'stringDecoder',
+    'console', 'os', 'fs', 'sd', 'tls','net', 'dns', 'url', 'util', 'path', 'zlib', 'http', 'https', 'dgram',
+    'stream', 'buffer', 'domain', 'crypto', 'events', 'punycode', 'readline', 'querystring',
 
     // Preinstalled modules
     'db',
     'con',
+    'common',
     'impress',
     'registry',
     'definition',