+++ deploy/debian-src.sh
@@ -15,15 +15,15 @@ else
 sudo apt-get -y install g++
 fi
 cd /usr/src
-wget http://nodejs.org/dist/v5.8.0/node-v5.8.0.tar.gz
-tar zxf node-v5.8.0.tar.gz
-rm -f ./node-v5.8.0.tar.gz
-cd node-v5.8.0
+wget http://nodejs.org/dist/v6.2.2/node-v6.2.2.tar.gz
+tar zxf node-v6.2.2.tar.gz
+rm -f ./node-v6.2.2.tar.gz
+cd node-v6.2.2
 ./configure
 make
 make install
 cd ~
-rm -rf /usr/src/node-v5.8.0
+rm -rf /usr/src/node-v6.2.2
 sudo apt-key adv --keyserver hkp://keyserver.ubuntu.com:80 --recv EA312927
 echo "deb http://repo.mongodb.org/apt/debian wheezy/mongodb-org/3.2 main" | sudo tee /etc/apt/sources.list.d/mongodb.list
 sudo apt-get -y update